+++ SRC/markSystem.cpp
@@ -496,13 +496,16 @@ static void SetSequenceStamp() //   mark words in sequence, original and canonic
 	//   consider all sets of up to 5-in-a-row 
 	for (int i = startSentence; i <= (int)endSentence; ++i)
 	{
-		if (!IsAlphaUTF8OrDigit(*wordStarts[i]) ) continue; // we only composite words, not punctuation or quoted stuff
-		if (IsDate(wordStarts[i])) // 1 word date
+		while (wordlist)
 		{
-			NextInferMark();
-			MarkFacts(0, true, MakeMeaning(FindWord((char*)"~dateinfo")), i, i, false, true);
-			continue;
+			int* chunk = (int*)Index2Stack(wordlist);
+			wordlist = chunk[0];
+			WORDP D = Index2Word(chunk[1]);
+			D->internalBits ^= BEEN_HERE;
 		}
+
+		if (!IsAlphaUTF8OrDigit(*wordStarts[i]) ) continue; // we only composite words, not punctuation or quoted stuff
+		if (IsDate(wordStarts[i])) continue;// 1 word date, caught later
 		// check for dates
 		int start,end;
 		if (DateZone(i,start,end) && i != wordCount)
@@ -673,12 +676,6 @@ void MarkAllImpliedWords()
 	for (i = 1; i <= wordCount; ++i)  capState[i] = IsUpperCase(*wordStarts[i]); // note cap state
 	failFired = false;
 	TagIt(); // pos tag and maybe parse
-	if (externalPostagger)
-	{
-		(*externalPostagger)();
-		startSentence = 1; // default when you dont know any better
-		endSentence = wordCount;
-	}
 
 	if ( prepareMode == POS_MODE || tmpPrepareMode == POS_MODE || prepareMode == PENN_MODE || prepareMode == POSVERIFY_MODE  || prepareMode == POSTIME_MODE ) 
 	{
@@ -777,8 +774,15 @@ void MarkAllImpliedWords()
 		{
 			if (IsDigit(*wordStarts[i]) && IsDigit(wordStarts[i][1])  && IsDigit(wordStarts[i][2]) && IsDigit(wordStarts[i][3])  && !wordStarts[i][4]) MarkFacts(0, ucase,MakeMeaning(FindWord((char*)"~yearnumber")),i,i);
 		}
+		if (IsDate(wordStarts[i]))
+		{
+			NextInferMark();
+			MarkFacts(0, true, MakeMeaning(FindWord((char*)"~dateinfo")), i, i, false, false);
+			MarkFacts(0, true, MakeMeaning(FindWord((char*)"~formatteddate")), i, i, false, false);
+		}
+
 		int number = IsNumber(wordStarts[i]);
-		if (number)
+		if (number && number != NOT_A_NUMBER)
 		{
 			if (!wordCanonical[i][1] || !wordCanonical[i][2]) // 2 digit or 1 digit
 			{