+++ changes.md
@@ -1,3 +1,11 @@
+# Version 7.53 8/12/2017
+1. :tsv  convert tsv to table form (putting quotes around things without it)
+2. %externaltagging
+3. improved control script manual
+
+# Version 7.52 7/8/2017
+1. ^respond(~xxx TEST) conditional execution of a topic to see if a match would occur (no output)
+2. revision of german noun pos tags in dictionary
 
 # Version 7.51  6/25/2017
 1. ^pos(isuppercase xx)  does it begin with uppercase letter